# backend/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Library Management API",
    "version": "0.0.1-SNAPSHOT",
    "title": "Spring Boot Rest API",
    "contact": {
      "name": "Ujeneza Divine",
      "url": "medium.com/codefountain",
      "email": "ujenezadivine@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8081",
  "basePath": "/",

  "paths": {
    "/books": {
      "get": {
        "tags": [
          "book-controller"
        ],
        "summary": "View all Books",
        "operationId": "getBooksUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all Books"
          
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "/books/:id":{
      "get": {
        "tags": [
          "book-controller"
        ],
        "summary": "Get a book information by id",
        "operationId": "getBooksByIdUsingGet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved a Book "
           
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "/auth/student/:id":{
      "delete": {
        "tags": ["student-controller"],
        "summary": "Deletes specific student with the supplied student id",
        "operationId": "deleteUsingDELETE",
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deletes the specific student"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "/auth/student-signin": {
      "post": {
        "tags": ["student-controller"],
        "summary": "Create a new student",
        "operationId": "createUserUsingPOST",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new Student"
       
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "/auth/student-login": {
      "post": {
        "tags": ["student-controller"],
        "summary": "Retrieve specific user to see if the have  to signin ",
        "operationId": "getUserUsingPost",
        "produces": ["*/*"],
        "parameters": [
            {
              "in": "body",
              "name": "user",
              "description": "user",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
        "responses": {
          "200": {
            "description": "Successfully logged in ",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
     
    }

   
  }
}